libsegment: + all in memory cache (backport trunk r73498)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_7_6@73500 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/segment/close.c
@@ -8,7 +8,7 @@
  *
  * \author GRASS GIS Development Team
  *
- * \date 2012
+ * \date 2018
  */
 
 #include <unistd.h>
@@ -35,12 +35,19 @@ int Segment_close(SEGMENT *SEG)
     if (SEG->open != 1)
 	return -1;
 
-    Segment_release(SEG);
-    close(SEG->fd);
-    unlink(SEG->fname);
+    if (SEG->scb) {
+	Segment_release(SEG);
+	close(SEG->fd);
+	unlink(SEG->fname);
 
-    SEG->fd = -1;
-    SEG->fname = NULL;
+	SEG->fd = -1;
+	SEG->fname = NULL;
+    }
+    else {
+	G_free(SEG->cache);
+    }
+
+    SEG->open = 0;
 
     return 1;
 }
